+++ services/marketplace/resource/model.go
@@ -23,22 +23,23 @@ func OfferingComponentType() types.ObjectType {
 	return types.ObjectType{AttrTypes: map[string]attr.Type{
 		"article_code":          types.StringType,
 		"billing_type":          types.StringType,
-		"default_limit":         types.Int64Type,
+		"default_limit":         types.Float64Type,
 		"description":           types.StringType,
 		"factor":                types.Int64Type,
 		"is_boolean":            types.BoolType,
 		"is_builtin":            types.BoolType,
 		"is_prepaid":            types.BoolType,
-		"limit_amount":          types.Int64Type,
+		"limit_amount":          types.Float64Type,
+		"limit_decimal_places":  types.Int64Type,
 		"limit_period":          types.StringType,
-		"max_available_limit":   types.Int64Type,
+		"max_available_limit":   types.Float64Type,
 		"max_prepaid_duration":  types.Int64Type,
 		"max_renewal_duration":  types.Int64Type,
-		"max_value":             types.Int64Type,
+		"max_value":             types.Float64Type,
 		"measured_unit":         types.StringType,
 		"min_prepaid_duration":  types.Int64Type,
 		"min_renewal_duration":  types.Int64Type,
-		"min_value":             types.Int64Type,
+		"min_value":             types.Float64Type,
 		"name":                  types.StringType,
 		"offering_uuid":         types.StringType,
 		"overage_component":     types.StringType,
@@ -86,7 +87,7 @@ func OrderInProgressType() types.ObjectType {
 		"error_updated_at":                      types.StringType,
 		"fixed_price":                           types.Float64Type,
 		"issue":                                 OrderInProgressIssueType(),
-		"limits":                                types.MapType{ElemType: types.Int64Type},
+		"limits":                                types.MapType{ElemType: types.Float64Type},
 		"marketplace_resource_uuid":             types.StringType,
 		"new_cost_estimate":                     types.StringType,
 		"new_plan_billing_mode":                 types.StringType,
@@ -508,11 +509,11 @@ func (model *MarketplaceResourceModel) CopyFrom(ctx context.Context, apiResp Mar
 	}
 
 	if apiResp.Limits != nil {
-		valLimits, diagsLimits := types.MapValueFrom(ctx, types.Int64Type, apiResp.Limits)
+		valLimits, diagsLimits := types.MapValueFrom(ctx, types.Float64Type, apiResp.Limits)
 		diags.Append(diagsLimits...)
 		model.Limits = valLimits
 	} else {
-		model.Limits = types.MapNull(types.Int64Type)
+		model.Limits = types.MapNull(types.Float64Type)
 	}
 
 	model.Name = common.StringPointerValue(apiResp.Name)